+++ cmd/cluster/cluster.go
@@ -68,6 +68,16 @@ var (
 		Version:  "v1alpha1",
 		Resource: "astartes",
 	}
+	aviV1Alpha1 = schema.GroupVersionResource{
+		Group:    "api.astarte-platform.org",
+		Version:  "v1alpha1",
+		Resource: "astartevoyageringresses",
+	}
+	adiV1Alpha1 = schema.GroupVersionResource{
+		Group:    "ingress.astarte-platform.org",
+		Version:  "v1alpha1",
+		Resource: "astartedefaultingresses",
+	}
 	crdResource = schema.GroupVersionResource{
 		Group:    "apiextensions.k8s.io",
 		Version:  "v1beta1",